# Double performance of rgb / rgba encode methods in BMPEncode