Original:
```
test bmp::encoder::tests::bench_encode_rgb    ... bench:   5,782,530 ns/iter (+/- 218,122)
test bmp::encoder::tests::bench_encode_rgba   ... bench:   8,998,140 ns/iter (+/- 16,662,076)
```

New:
```
test bmp::encoder::tests::bench_encode_rgb    ... bench:   3,316,475 ns/iter (+/- 209,397)
test bmp::encoder::tests::bench_encode_rgba   ... bench:   4,440,780 ns/iter (+/- 436,233)
```

For some reason, the rgba benchmark seems to have a large +/- of ~17M ns, but the actual timings are always ~9M ns.

I've included benchmarks, and I've tested against the BMP tests, which seem to work fine.