@philippeitis 2020-02-18
I've made the requested changes.
I also simplified encode_gray (writing with a buffer of values and using the reverse iterator).

An additional optimization which could be particularly effective is writing to an intermediate buffer for each row to avoid calling write_all() / write_u8 in a tight loop, but this depends on the size of the buffer, and could be abused to allocate excessively large buffers.