@HeroicKatora 2020-02-18
@philippeitis An intermediate buffer is already possible by using a `BufWriter`, there's shouldn't be a need to worry about it in a structure that's generic over the underlying writer. Maybe that means that we should benchmark over different writers, to get a more representative picture.