@philippeitis 2020-02-18
Given that the writer used in the benchmarks is a raw vector (to avoid filesystem effects on writing), I think the speedup comes from reducing the number of calls in tight loops - each call to write_u8()? or write_all()? introduces additional overhead (simply calling the function, boundary checks, checking that a write is successful) which can be condensed into a single write at the end of the loop. A `bufWriter` would help in the case that we're writing to a filesystem, which is not what I am testing in the benchmarks.