@philippeitis 2020-02-18
```
```
```   fn encode_gray(
        &mut self,
        image: &[u8],
        width: u32,
        height: u32,
        row_pad_size: u32,
        bytes_per_pixel: u32,
    ) -> io::Result<()> {
        // write grayscale palette
        for val in 0u8..=255 {
            // each color is written as BGRA, where A is always 0 and since only grayscale is being written, B = G = R = index
            self.writer.write_all(&[val, val, val, 0])?;
        }

        // write image data
        let width = width as usize;
        let height = height as usize;
        let x_stride = bytes_per_pixel as usize;
        let y_stride = width * x_stride;
        let mut buf = Vec::with_capacity(width);
        for row in (0..height).rev() {
            // from the bottom up
            let row_start = row * y_stride;
            for col in 0..width {
                let pixel_start = (row_start + (col * x_stride)) as usize;
                // color value is equal to the palette index
                buf.push(image[pixel_start]);
            }
            self.writer.write_all(&buf)?;
            buf.clear();

            self.write_row_pad(row_pad_size)?;
        }

        Ok(())
    }
```

This runs in
```
test bmp::encoder::tests::bench_encode_gray   ... bench:   1,970,567 ns/iter (+/- 1,026,174)
```
vs. the previous version's
```
test bmp::encoder::tests::bench_encode_gray   ... bench:   4,316,750 ns/iter (+/- 567,438)
```

(The large benchmark times are an artifact of my laptop's clock performance - for comparision, when underclocked, the original encode_rgba function runs in ~18M ns and the new one runs in ~7.5M ns, while the overclocked times are as in the original pull comment.)