@philippeitis 2020-02-20
Alternative, if we want to ensure that this intermediate buffer never exceeds a certain capacity, we could use this implementation:
```
    fn encode_gray(
        &mut self,
        image: &[u8],
        width: u32,
        height: u32,
        row_pad_size: u32,
        bytes_per_pixel: u32,
    ) -> io::Result<()> {
        // write grayscale palette
        for val in 0u8..=255 {
            // each color is written as BGRA, where A is always 0 and since only grayscale is being written, B = G = R = index
            self.writer.write_all(&[val, val, val, 0])?;
        }
        // write image data
        let max_width = 4096usize;
        let width = width as usize;
        let height = height as usize;
        let x_stride = bytes_per_pixel as usize;
        let y_stride = width * x_stride;
        let buf_width = if width < max_width {width} else {max_width};
        let mut buf = vec![0u8; buf_width];
        // from the bottom up
        for row in (0..height).rev() {
            let mut col = 0;
            let mut pos = row * y_stride;
            let mut num_elems = 0;
            while col < width {
                while num_elems < buf_width && col < width {
                    buf[num_elems] = image[pos];
                    num_elems += 1;
                    pos += x_stride;
                    col += 1;
                }
                self.writer.write_all(&buf[..num_elems])?;
                num_elems = 0;
            }
            self.write_row_pad(row_pad_size)?;
        }

        Ok(())
    }
```

Old:
```
test bmp::encoder::tests::bench_encode_gray   ... bench:   5,069,330 ns/iter (+/- 453,906)
```
vs. new
```
test bmp::encoder::tests::bench_encode_gray   ... bench:   1,425,107 ns/iter (+/- 210,943)
```