@HeroicKatora 2020-02-20
That's *very* curious because it is precisely what `BufWriter` should provide. Have you *tried* using a `BufWriter<Vec<u8>>` as the output writer of the benchmark, either by replacing the type in the benchmark or by wrapping the provided writer within `encode_gray`? It seems plausible that IO performance would also in general improve from grouping writes into larger ones but very weird for in-memory IO. We really should be careful to not optimize with a bad metric at the cost of complicating the logic and inhibiting future optimization.