@philippeitis 2020-02-20
Benchmark times (best of 5 in both cases)
Old:
```
test bmp::encoder::tests::bench_encode_gray          ... bench:   2,700,717 ns/iter (+/- 209,294)
test bmp::encoder::tests::bench_encode_gray_buf      ... bench:   7,854,635 ns/iter (+/- 1,675,822)
test bmp::encoder::tests::bench_encode_gray_buf_file ... bench:  15,644,970 ns/iter (+/- 2,933,322)
```
New:
```
test bmp::encoder::tests::bench_encode_gray          ... bench:   2,058,517 ns/iter (+/- 238,720)
test bmp::encoder::tests::bench_encode_gray_buf      ... bench:   1,440,242 ns/iter (+/- 195,000)
test bmp::encoder::tests::bench_encode_gray_buf_file ... bench:   8,846,160 ns/iter (+/- 1,525,621)
```

The buffered writer does seem to help, in the case that we also buffer our own writes to it, which was quite surprising. It seems that using this intermediate buffer is effective for most writers..