@philippeitis 2020-02-20
I've included the benchmarks (without the implementation - which is as it appears in the comments), but have left the file benchmarks commented out as I'm not sure how you'd want to handle the temporary file creation and deletion.